# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Work Manager API
  version: 1.0.0
  description: |
    The two token-authenticated REST surfaces of this application. Everything here is **verified against the
    running code on every test run** (`tests/Functional/Api/OpenApiContractTest.php`): each operation is called
    for real and its status codes, headers and JSON bodies are validated against the schemas below, and
    `OpenApiSpecTest` fails if a route exists that is not documented (or the reverse). If this file and the code
    disagree, the build is red — fix whichever one is wrong.

    ## Authentication — both APIs

    Both surfaces use the same **personal access tokens** (`auth-pat-bundle`): create one at `/account/tokens` and
    send it as `Authorization: Bearer <token>`. The plaintext is shown once; only its hash is stored. A revoked,
    expired or unknown token, or one whose owner is deactivated, is answered `401`. What a token may do follows its
    owner's roles (ADR-068): the Admin API needs an account holding an admin role (else `403`); some endpoints need
    tech support. Without `auth-pat-bundle` no token can authenticate, so both APIs refuse every call (`401`).

    ## Admin API — `/admin-api/*`

    Account and instance management, for accounts holding `ROLE_ADMIN` or above. An admin manages regular users;
    a super admin also manages admins; tech-support accounts are visible only to tech support.

...

    ## User API — `/api/*`

...

...

...
servers:
...
paths:
...
  /admin-api/users/{id}:
    parameters:
      - $ref: '#/components/parameters/UserId'
    get:
      operationId: getUser
      tags: [Admin users]
      summary: Get a user
      security:
        - adminBearer: []
      responses:
        '200':
          description: The user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/UserNotFound' }

    patch:
      operationId: updateUser
      tags: [Admin users]
      summary: Update a user
      description: |
        PATCH semantics — only the keys you send are validated and applied (`email`, `name`, `role`, `status`).
        Setting `status` to `inactive` cuts access immediately (sessions, tokens and recovery links are revoked).
      security:
        - adminBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateUserRequest' }
            example:
              name: Jane Q. Doe
      responses:
        '200':
          description: The updated user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/UserNotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }

    delete:
      operationId: deleteUser
      tags: [Admin users]
      summary: Delete (deactivate) a user
      description: Soft delete — sets `status` to `inactive` and revokes outstanding recovery tokens. The record is kept.
      security:
        - adminBearer: []
      responses:
        '204': { description: Deleted (deactivated). No body. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/UserNotFound' }
...
  /admin-api/htaccess-lock:
    get:
      operationId: getHtaccessLock
      x-audience: tech-support
      tags: [Admin htaccess lock]
      summary: Get the Htaccess Lock
      description: The current policy, the health of the `.htaccess` file, and the caller's IP as the server sees it.
      security:
        - adminBearer: []
      responses:
        '200':
          description: The lock.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HtaccessLock' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    patch:
      operationId: updateHtaccessLock
      x-audience: tech-support
      tags: [Admin htaccess lock]
      summary: Change the lock's settings
      description: |
        PATCH semantics — only the keys you send change. `ips` and `exempt_paths` **replace** the whole list (use
        the `/ips` and `/exempt-paths` calls to add or remove one entry). Whether the lock is on is not patched
        here: use `/enable` and `/disable`. When the lock is on, the result must still whitelist the caller's own IP.
      security:
        - adminBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateHtaccessLockRequest' }
            example:
              status_code: 404
              exempt_paths: ['/health']
      responses:
        '200':
          description: The lock after the change.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HtaccessLock' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/HtaccessLockRejected' }
        '500': { $ref: '#/components/responses/HtaccessWriteFailed' }
  /admin-api/htaccess-lock/enable:
    post:
      operationId: enableHtaccessLock
      x-audience: tech-support
      tags: [Admin htaccess lock]
      summary: Switch the lock on
      description: |
        Writes the managed block into `.htaccess`. Refused (422) when the whitelist is empty or does not cover the
        caller's own IP — that would lock the caller out. Allow a couple of seconds for the web server to notice.
        Run the self-test afterwards: a server that ignores `.htaccess` fails open.
      security:
        - adminBearer: []
      responses:
        '200':
          description: The lock, now on.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HtaccessLock' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/HtaccessLockRejected' }
        '500': { $ref: '#/components/responses/HtaccessWriteFailed' }
  /admin-api/htaccess-lock/disable:
    post:
      operationId: disableHtaccessLock
      x-audience: tech-support
      tags: [Admin htaccess lock]
      summary: Switch the lock off
      description: Removes the managed block from `.htaccess`; the saved whitelist is kept. Never refused for validation reasons.
      security:
        - adminBearer: []
      responses:
        '200':
          description: The lock, now off.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HtaccessLock' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/HtaccessWriteFailed' }
  /admin-api/htaccess-lock/ips:
    get:
      operationId: listWhitelistedIps
      x-audience: tech-support
      tags: [Admin htaccess lock]
      summary: List whitelisted IPs
      security:
        - adminBearer: []
      responses:
        '200':
          description: The whitelist (canonical IPv4/IPv6 addresses and CIDR ranges).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EntryList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      operationId: addWhitelistedIp
      x-audience: tech-support
      tags: [Admin htaccess lock]
      summary: Add a whitelisted IP
      description: |
        Accepts an IPv4/IPv6 address or a CIDR range (a `/0` is refused: it would let everyone in). The stored
        form is canonical, so `2001:DB8::1` and `2001:db8:0:0:0:0:0:1` are the same entry. If the lock is on the
        change is applied to `.htaccess` immediately.
      security:
        - adminBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddIpRequest' }
            example: { ip: 203.0.113.0/24 }
      responses:
        '201':
          description: The whitelist after the addition.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EntryList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: That IP is already whitelisted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: 203.0.113.0/24 is already whitelisted. }
        '422': { $ref: '#/components/responses/HtaccessLockRejected' }
        '500': { $ref: '#/components/responses/HtaccessWriteFailed' }

    delete:
      operationId: removeWhitelistedIp
      x-audience: tech-support
      tags: [Admin htaccess lock]
      summary: Remove a whitelisted IP
      description: |
        The IP goes in the `ip` query parameter (CIDR slashes are fine there). Removing the entry that covers the
        caller's own IP while the lock is on is refused (422) — it would lock the caller out; so is emptying the
        whitelist while the lock is on.
      security:
        - adminBearer: []
      parameters:
        - name: ip
          in: query
          required: true
          description: The address or CIDR to remove, as listed.
          schema: { type: string, minLength: 1 }
          example: 203.0.113.0/24
      responses:
        '200':
          description: The whitelist after the removal.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EntryList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: That IP is not whitelisted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: 203.0.113.0/24 is not whitelisted. }
        '422': { $ref: '#/components/responses/HtaccessLockRejected' }
        '500': { $ref: '#/components/responses/HtaccessWriteFailed' }
  /admin-api/htaccess-lock/exempt-paths:
    get:
      operationId: listExemptPaths
      x-audience: tech-support
      tags: [Admin htaccess lock]
      summary: List exempt paths
      description: URL paths reachable from any IP while the lock is on (e.g. a health check or an inbound webhook).
      security:
        - adminBearer: []
      responses:
        '200':
          description: The exempt paths.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EntryList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      operationId: addExemptPath
      x-audience: tech-support
      tags: [Admin htaccess lock]
      summary: Add an exempt path
      description: A URL path starting with `/` (no domain); a trailing `*` makes it a prefix. `/*` is refused.
      security:
        - adminBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddExemptPathRequest' }
            example: { path: /health }
      responses:
        '201':
          description: The exempt paths after the addition.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EntryList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: That path is already exempt.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: /health is already exempt. }
        '422': { $ref: '#/components/responses/HtaccessLockRejected' }
        '500': { $ref: '#/components/responses/HtaccessWriteFailed' }

    delete:
      operationId: removeExemptPath
      x-audience: tech-support
      tags: [Admin htaccess lock]
      summary: Remove an exempt path
      security:
        - adminBearer: []
      parameters:
        - name: path
          in: query
          required: true
          description: The path to remove, exactly as listed.
          schema: { type: string, minLength: 1 }
          example: /health
      responses:
        '200':
          description: The exempt paths after the removal.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EntryList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: That path is not exempt.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: /health is not exempt. }
        '422': { $ref: '#/components/responses/HtaccessLockRejected' }
        '500': { $ref: '#/components/responses/HtaccessWriteFailed' }
  /admin-api/htaccess-lock/self-test:
    get:
      operationId: getHtaccessLockSelfTest
      x-audience: tech-support
      tags: [Admin htaccess lock]
      summary: Get the last self-test result
      security:
        - adminBearer: []
      responses:
        '200':
          description: The most recent enforcement self-test.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SelfTestReport' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: The self-test has never been run.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: The self-test has never been run. }

    post:
      operationId: runHtaccessLockSelfTest
      x-audience: tech-support
      tags: [Admin htaccess lock]
      summary: Run the enforcement self-test
      description: |
        Proves against the real web server that the lock is enforced (a server that ignores `.htaccess` fails
        open, silently). Briefly swaps temporary blocks into `.htaccess` and requests the site over loopback, then
        restores the file byte-for-byte. **Real visitors are blocked for a few seconds while it runs**, and the
        call takes that long to answer. A failing test is still `200` — read `passed`.
      security:
        - adminBearer: []
      responses:
        '200':
          description: The test report (also stored as the "last result").
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SelfTestReport' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    adminBearer:
      type: http
      scheme: bearer
      description: A personal access token (created at `/account/tokens`) whose owner holds an admin role.
    userBearer:
      type: http
      scheme: bearer
      description: A user's personal access token, created at `/account/tokens`.
  parameters:
    UserId:
      name: id
      in: path
      required: true
      description: User id.
      schema: { type: integer, minimum: 1 }
      example: 7
  responses:
    Unauthorized:
      description: The bearer token is missing, invalid, revoked, expired, or belongs to an inactive account.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: Missing or invalid Authorization header. }
    UserNotFound:
      description: No user with that id.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: User not found. }
    ValidationFailed:
      description: One or more fields failed validation. Keys are field names.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationErrors' }
          example:
            errors:
              email: This email address is already registered.
    Forbidden:
      description: The token is valid but its admin is not allowed to call this endpoint (tech support only).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: Access denied. }
    HtaccessLockRejected:
      description: The request was refused and nothing was changed (bad input, or it would lock the caller out). `error` is the first reason; `errors` lists all of them.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/HtaccessLockRejected' }
          example:
            error: Your current IP (198.51.100.4) is not covered by the whitelist. Add it first, or this save would lock you out of the admin panel.
            errors:
              - Your current IP (198.51.100.4) is not covered by the whitelist. Add it first, or this save would lock you out of the admin panel.
    HtaccessWriteFailed:
      description: The input was fine but `.htaccess` could not be written (permissions, or hand-edited markers). Nothing was recorded as changed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'Could not update .htaccess: /var/www/public/.htaccess has unbalanced Htaccess Lock markers; refusing to edit it.' }
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: { type: string, minLength: 1 }
    ValidationErrors:
      type: object
      additionalProperties: false
      required: [errors]
      properties:
        errors:
          type: object
          minProperties: 1
          additionalProperties: { type: string, minLength: 1 }
...
    AccountStatus:
      type: string
      enum: [active, inactive]
    User:
      type: object
      additionalProperties: false
      required: [id, email, name, status, roles, created_at]
      properties:
        id: { type: integer, minimum: 1 }
        email: { type: string, format: email }
        name: { type: string }
        status: { $ref: '#/components/schemas/AccountStatus' }
        roles:
          type: array
          minItems: 1
          items: { type: string, enum: [ROLE_USER] }
        created_at: { type: string, format: date-time }
...
    UpdateUserRequest:
      type: object
      minProperties: 1
      properties:
        email: { type: string, format: email }
        name: { type: string, minLength: 1 }
        role: { type: string, enum: [ROLE_USER] }
        status: { $ref: '#/components/schemas/AccountStatus' }
...
    HtaccessLock:
      type: object
      additionalProperties: false
      required: [enabled, status_code, error_file, ips, exempt_paths, file, your_ip]
      properties:
        enabled: { type: boolean }
        status_code: { type: integer, enum: [403, 404], description: What blocked visitors get. 404 hides that the site is locked. }
        error_file: { type: string, description: 'URL path of the page shown to blocked visitors; "" = empty body.' }
        ips: { type: array, items: { type: string, minLength: 1 } }
        exempt_paths: { type: array, items: { type: string, minLength: 1 } }
        file:
          type: object
          additionalProperties: false
          required: [exists, writable, has_block, in_sync, error]
          description: Whether `.htaccess` agrees with the policy above (`in_sync`), and why not if it cannot be read.
          properties:
            exists: { type: boolean }
            writable: { type: boolean }
            has_block: { type: boolean }
            in_sync: { type: boolean }
            error: { type: [string, 'null'] }
        your_ip: { type: [string, 'null'], description: The caller's IP as the server sees it. }
    UpdateHtaccessLockRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        ips: { type: array, items: { type: string, minLength: 1 } }
        exempt_paths: { type: array, items: { type: string, minLength: 1 } }
        status_code: { type: integer, enum: [403, 404] }
        error_file: { type: string }
    AddIpRequest:
      type: object
      additionalProperties: false
      required: [ip]
      properties:
        ip: { type: string, minLength: 1 }
    AddExemptPathRequest:
      type: object
      additionalProperties: false
      required: [path]
      properties:
        path: { type: string, minLength: 1 }
    EntryList:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data: { type: array, items: { type: string, minLength: 1 } }
    HtaccessLockRejected:
      type: object
      additionalProperties: false
      required: [error, errors]
      properties:
        error: { type: string, minLength: 1 }
        errors: { type: array, minItems: 1, items: { type: string, minLength: 1 } }
    SelfTestReport:
      type: object
      additionalProperties: false
      required: [passed, ran_at, steps]
      properties:
        passed: { type: boolean }
        ran_at: { type: string, format: date-time }
        steps:
          type: array
          minItems: 1
          items:
            type: object
            additionalProperties: false
            required: [label, passed, detail]
            properties:
              label: { type: string }
              passed: { type: boolean }
              detail: { type: string }